Update the exchanger slice
